+++ js/data.js
@@ -1,0 +1,3 @@
+import {generateObjects} from './create-data.js';
+
+generateObjects();
